Atom and Rss 2.0 Feed generation via Attributes (no hard coding). Updated Tests (generating from the properties changes the order of the output). Changed Enclosure.Length to Int64 (from Int32). Added (optional) Indent property to Generators (beautify xml ouput). Added HomePage proprty to Author. Corrected spelling mistake in Image. Added Icon, WebMaster, ManagingEditor and TimeToLive properties to Feed. Lots of other changes and improvements.

+++ src/Enclosure.cs
@@ -6,7 +6,7 @@ namespace dng.Syndication
     {
         public Enclosure(
             string url,
-            int length,
+            long length,
             string mediaType)
         {
             if (string.IsNullOrWhiteSpace(url) || string.IsNullOrWhiteSpace(mediaType) || length < 0)
@@ -25,7 +25,7 @@ namespace dng.Syndication
         /// <summary>
         /// How big it is in bytes
         /// </summary>
-        public int Length { get; }
+        public long Length { get; }
 
         /// <summary>
         /// Standard MIME type
